+++ examples/default/deploy.tf
@@ -12,8 +12,15 @@ locals {
   }
 }
 
+resource "aws_iam_policy" "extra_policy" {
+  name        = format("%sExtraPolicy", random_string.prefix.result)
+  description = "extra policy"
+  path        = "/"
+  policy      = data.aws_iam_policy_document.extra_policy.json
+}
+
 module "compute_environment" {
-  source = "git::https://scm.dazzlingwrench.fxinnovation.com/fxinnovation-public/terraform-module-aws-batch.git?ref=0.1.0"
+  source = "git::https://scm.dazzlingwrench.fxinnovation.com/fxinnovation-public/terraform-module-aws-batch.git?ref=0.1.1"
 
   prefix = format("tft%s-", random_string.prefix.result)
 
@@ -28,15 +35,16 @@ module "default" {
 
   prefix = format("tft%s-", random_string.prefix.result)
 
-  name                = "batchjob"
-  job_queue_arn       = module.compute_environment.batch_job_queue_this_arn
-  schedule_expression = "cron(0 0 * * ? 1970)"
-  properties = jsonencode({
+  name                           = "batchjob"
+  job_queue_arn                  = module.compute_environment.batch_job_queue_this_arn
+  execution_role_extras_policies = [aws_iam_policy.extra_policy.arn]
+  schedule_expression            = "cron(0 0 * * ? 1970)"
+  properties = {
     "image" : "busybox",
     "command" : ["ls", "-la"],
     "memory" : 256,
     "vcpus" : 1,
-  })
+  }
 
   tags = local.tags
 }